+++ daiv/automation/agents/image_url_extractor/conf.py
@@ -12,8 +12,8 @@ class ImageURLExtractorSettings(BaseSettings):
     model_config = SettingsConfigDict(secrets_dir="/run/secrets", env_prefix="IMAGE_URL_EXTRACTOR_")
 
     NAME: str = Field(default="ImageURLExtractor", description="Name of the image URL extractor agent.")
-    MODEL_NAME: ModelName = Field(
-        default=ModelName.GPT_4O_MINI, description="Model name to be used for image URL extractor."
+    MODEL_NAME: ModelName | str = Field(
+        default=ModelName.GEMINI_2_0_FLASH_LITE, description="Model name to be used for image URL extractor."
     )