+++ daiv/automation/agents/issue_addressor/conf.py
@@ -13,11 +13,8 @@ class IssueAddressorSettings(BaseSettings):
 
     NAME: str = Field(default="IssueAddressor", description="Name of the issue addressor agent.")
     RECURSION_LIMIT: int = Field(default=50, description="Recursion limit for the issue addressor agent.")
-    ASSESSMENT_MODEL_NAME: ModelName = Field(
-        default=ModelName.GPT_4O_MINI, description="Model name to be used for issue assessment."
-    )
-    FALLBACK_ASSESSMENT_MODEL_NAME: ModelName = Field(
-        default=ModelName.CLAUDE_3_5_HAIKU, description="Fallback model name to be used for issue assessment."
+    ASSESSMENT_MODEL_NAME: ModelName | str = Field(
+        default=ModelName.GEMINI_2_0_FLASH, description="Model name to be used for issue assessment."
     )