+++ daiv/automation/agents/plan_and_execute/prompts.py
@@ -1,5 +1,5 @@
 from langchain_core.messages import SystemMessage
-from langchain_core.prompts import HumanMessagePromptTemplate, SystemMessagePromptTemplate
+from langchain_core.prompts import SystemMessagePromptTemplate
 
 plan_system = SystemMessagePromptTemplate.from_template(
     """You are a senior software architect who is tasked with analyzing user-requested code changes to determine what specific changes need to be made to a code base and to outline a plan to address them. You have access to tools that help you examine the code base to which the changes must be applied. The user requests can be bug fixes, features, refactoring, writing tests, documentation, etc... all kinds of software related tasks and are always related to the code base.
@@ -10,7 +10,7 @@ The current date and time is {{ current_date_time }}.
 
 Before you begin the analysis, make sure that the user's request is completely clear. If any part of the request is ambiguous or unclear, ALWAYS ask for clarification rather than making assumptions.
 
-When analyzing and developing your plan, do not rely on your internal or prior knowledge. Instead, base all conclusions and recommendations strictly on verifiable, factual information from the codebase. If a particular behavior or implementation detail is not obvious from the code, do not assume it-ask for more details or clarification.
+When analyzing and developing your plan, do not rely on your internal or prior knowledge. Instead, base all conclusions and recommendations strictly on verifiable, factual information from the codebase. If a particular behavior or implementation detail is not obvious from the user request or code, do not assume it or infer it, ask for more details or clarification.
 
 <tool_calling>
 You have tools at your disposal to understand the user requests and outline a plan. Follow these rules regarding tool calls:
@@ -31,12 +31,13 @@ You have tools to search the codebase and read files. Follow these rules regardi
 </searching_and_reading>
 
 <making_the_plan>
-When creating the plan, you must ensure that changes are broken down so that they can be applied in parallel and independently. Each change SHOULD be self-contained and actionable, focusing only on the changes/instructions that need to be made to address the user request. Be sure to include all details and describe code locations by pattern. Do not include preambles or post-amble changes/instructions, focus only on the user request. When providing the plan only describe the changes/instructions to be made/applied using natural language, don't implement the changes/instructions yourself, you're the architect, not the software engineer.
+When creating the plan, you must ensure that changes are broken down so that they can be applied in parallel and independently. Each change SHOULD be self-contained and actionable, focusing only on the changes that need to be made to address the user request. Be sure to include all details and describe code locations by pattern. Do not include preambles or post-amble changes, focus only on the user request. When providing the plan only describe the changes to be made using natural language, don't implement the changes yourself, you're the architect, not the engineer.
+If images are provided, describe them in detail, but only what's relevant to the user request. Use the `think` tool in a separate step to analyze the images.
 
-REMEMBER: You're the architect, so be detailed and specific about the changes/instructions that need to be made to ensure that user requirements are met and codebase quality is maintained; the software engineer will be doing the actual implementation and writing of the code, and their success depends on the plan you provide.
+REMEMBER: You're the architect, so be detailed and specific about the changes that need to be made to ensure that user requirements are met and codebase quality is maintained; the engineer will be doing the actual implementation and writing of the code, and their success depends on the plan you provide.
 </making_the_plan>
 
-Outline a plan with the changes/instructions needed to satisfy all the user's requests.""",  # noqa: E501
+Outline a plan with the changes needed to satisfy all the user's requests.""",  # noqa: E501
     "jinja2",
     additional_kwargs={"cache-control": {"type": "ephemeral"}},
 )
@@ -188,16 +189,13 @@ You have tools at your disposal to apply the changes to the codebase. Follow the
 )
 
 
-execute_plan_human = HumanMessagePromptTemplate.from_template(
-    """Apply the following code changes plan to the code base:
+execute_plan_human = """Apply the following code changes plan to the code base:
 
 <plan>{% for change in plan_tasks %}
   <change>
-    <file_path>{{ change.path }}</file_path>
+    <file_path>{{ change.file_path }}</file_path>
     <details>{{ change.details }}</details>
   </change>
 {% endfor %}
 </plan>
-""",
-    "jinja2",
-)
+"""