+++ daiv/automation/agents/pr_describer/conf.py
@@ -8,11 +8,8 @@ class PRDescriberSettings(BaseSettings):
     model_config = SettingsConfigDict(secrets_dir="/run/secrets", env_prefix="PR_DESCRIBER_")
 
     NAME: str = Field(default="PullRequestDescriber", description="Name of the PR describer agent.")
-    MODEL_NAME: ModelName = Field(
-        default=ModelName.CLAUDE_3_5_HAIKU, description="Model name to be used for PR describer."
-    )
-    FALLBACK_MODEL_NAME: ModelName = Field(
-        default=ModelName.GPT_4O_MINI, description="Fallback model name to be used for PR describer."
+    MODEL_NAME: ModelName | str = Field(
+        default=ModelName.GEMINI_2_0_FLASH_LITE, description="Model name to be used for PR describer."
     )